Fix build without PCH.
